Fix Eslint Flat Config Code
When you try to run the config, not only does this code not work but typescript starts to yell at you. The solution was to take out 'eslint:recommended' and I changed it to the example in the official eslint docs https://eslint.org/docs/latest/use/configure/migration-guide#predefined-and-shareable-configs.